refactor: move some of the direction removal code to a composable module
